Use existing ruby for local dev
getruby.cmd depends on curl, which isn't on a windows machine by default.  Rather than check if curl.exe is in the path, then install that, then install ruby, I just check that ruby's in the path.  If it is, then move along.

It means that ruby is a pre-req for local dev.  But you probably knew that and it'll be on the local machine if you're building Jekyll.